requirements.txt
Django
openai
python-dotenv
